+++ packages/terra-list/src/List.jsx
@@ -245,7 +245,7 @@ const List = ({
   const getStyleforDrag = (ListItem, snapshot, provider) => {
     const styleProperties = provider?.draggableProps?.style;
     if (styleProperties && snapshot && snapshot.isDragging) {
-      styleProperties['z-index'] = ListItem?.props?.zIndex ? `${ListItem.props.zIndex}` : '6001';
+      styleProperties.zIndex = ListItem?.props?.zIndex ? `${ListItem.props.zIndex}` : '6001';
     }
     return styleProperties;
   };